buildstatus/requirements.txt
gpiozero
rpi.gpio
jenkins-webapi
sshtunnel
